move PolicyMember to data/types/
