+++ src/components/SelectAccountablePerson.svelte
@@ -3,7 +3,8 @@ import DependentForm from './forms/DependentForm.svelte'
 import { AccountablePersonOptions, selectedAccountablePersonOptions } from 'data/accountablePersons'
 import { addDependent, dependentsByPolicyId, initialized, PolicyDependent } from 'data/dependents'
 import { policies, Policy, PolicyType } from 'data/policies'
-import { membersByPolicyId, PolicyMember, selectedPolicyMembers } from 'data/policy-members'
+import { membersByPolicyId, selectedPolicyMembers } from 'data/policy-members'
+import type { PolicyMember } from 'data/types/policy-members'
 import { Dialog, Select } from '@silintl/ui-components'
 import { createEventDispatcher } from 'svelte'
