+++ src/data/types/policy-members.ts
@@ -1,0 +1,9 @@
+export type PolicyMember = {
+  email: string
+  email_override: string
+  first_name: string
+  id: string
+  last_login_utc: string /*Date*/
+  last_name: string
+  country: string
+}
